# Improve Parquet Dictionary Handling