--- a/pkg/parquetquery/predicates.go
+++ b/pkg/parquetquery/predicates.go
@@ -20,6 +20,8 @@ type Predicate interface {
 // Case sensitive exact byte matching
 type StringInPredicate struct {
 	ss [][]byte
+
+	helper DictionaryPredicateHelper
 }
 
 var _ Predicate = (*StringInPredicate)(nil)
@@ -35,6 +37,8 @@ func NewStringInPredicate(ss []string) Predicate {
 }
 
 func (p *StringInPredicate) KeepColumnChunk(cc pq.ColumnChunk) bool {
+	p.helper.setNewRowGroup()
+
 	if ci := cc.ColumnIndex(); ci != nil {
 
 		for _, subs := range p.ss {
@@ -65,26 +69,28 @@ func (p *StringInPredicate) KeepValue(v pq.Value) bool {
 func (p *StringInPredicate) KeepPage(page pq.Page) bool {
 	// todo: check bounds
 
-	// If a dictionary column then ensure at least one matching
-	// value exists in the dictionary
-	dict := page.Dictionary()
-	if dict != nil && dict.Len() > 0 {
-		len := dict.Len()
+	return p.helper.keepPage(func() bool {
+		// If a dictionary column then ensure at least one matching
+		// value exists in the dictionary
+		dict := page.Dictionary()
+		if dict != nil && dict.Len() > 0 {
+			len := dict.Len()
 
-		for i := 0; i < len; i++ {
-			dictionaryEntry := dict.Index(int32(i)).ByteArray()
-			for _, subs := range p.ss {
-				if bytes.Equal(dictionaryEntry, subs) {
-					// At least 1 string present in this page
-					return true
+			for i := 0; i < len; i++ {
+				dictionaryEntry := dict.Index(int32(i)).ByteArray()
+				for _, subs := range p.ss {
+					if bytes.Equal(dictionaryEntry, subs) {
+						// At least 1 string present in this page
+						return true
+					}
 				}
 			}
+
+			return false
 		}
 
-		return false
-	}
-
-	return true
+		return true
+	})
 }
 
 // RegexInPredicate checks for match against any of the given regexs.
@@ -92,6 +98,8 @@ func (p *StringInPredicate) KeepPage(page pq.Page) bool {
 type RegexInPredicate struct {
 	regs    []*regexp.Regexp
 	matches map[string]bool
+
+	helper DictionaryPredicateHelper
 }
 
 var _ Predicate = (*RegexInPredicate)(nil)
@@ -135,6 +143,8 @@ func (p *RegexInPredicate) keep(v *pq.Value) bool {
 }
 
 func (p *RegexInPredicate) KeepColumnChunk(cc pq.ColumnChunk) bool {
+	p.helper.setNewRowGroup()
+
 	// Reset match cache on each row group change
 	p.matches = make(map[string]bool, len(p.matches))
 
@@ -148,29 +158,33 @@ func (p *RegexInPredicate) KeepValue(v pq.Value) bool {
 
 func (p *RegexInPredicate) KeepPage(page pq.Page) bool {
 
-	// If a dictionary column then ensure at least one matching
-	// value exists in the dictionary
-	dict := page.Dictionary()
-	if dict != nil && dict.Len() > 0 {
-		len := dict.Len()
+	return p.helper.keepPage(func() bool {
+		// If a dictionary column then ensure at least one matching
+		// value exists in the dictionary
+		dict := page.Dictionary()
+		if dict != nil && dict.Len() > 0 {
+			len := dict.Len()
 
-		for i := 0; i < len; i++ {
-			dictionaryEntry := dict.Index(int32(i))
-			if p.keep(&dictionaryEntry) {
-				// At least 1 dictionary entry matches
-				return true
+			for i := 0; i < len; i++ {
+				dictionaryEntry := dict.Index(int32(i))
+				if p.keep(&dictionaryEntry) {
+					// At least 1 dictionary entry matches
+					return true
+				}
 			}
+
+			return false
 		}
 
-		return false
-	}
-
-	return true
+		return true
+	})
 }
 
 type SubstringPredicate struct {
 	substring string
 	matches   map[string]bool
+
+	helper DictionaryPredicateHelper
 }
 
 var _ Predicate = (*SubstringPredicate)(nil)
@@ -204,22 +218,23 @@ func (p *SubstringPredicate) KeepValue(v pq.Value) bool {
 }
 
 func (p *SubstringPredicate) KeepPage(page pq.Page) bool {
-
-	// If a dictionary column then ensure at least one matching
-	// value exists in the dictionary
-	dict := page.Dictionary()
-	if dict != nil && dict.Len() > 0 {
-		len := dict.Len()
-		for i := 0; i < len; i++ {
-			if p.KeepValue(dict.Index(int32(i))) {
-				return true
+	return p.helper.keepPage(func() bool {
+		// If a dictionary column then ensure at least one matching
+		// value exists in the dictionary
+		dict := page.Dictionary()
+		if dict != nil && dict.Len() > 0 {
+			len := dict.Len()
+			for i := 0; i < len; i++ {
+				if p.KeepValue(dict.Index(int32(i))) {
+					return true
+				}
 			}
+
+			return false
 		}
 
-		return false
-	}
-
-	return true
+		return true
+	})
 }
 
 // IntBetweenPredicate checks for int between the bounds [min,max] inclusive
@@ -270,15 +285,19 @@ type GenericPredicate[T any] struct {
 	Fn      func(T) bool
 	RangeFn func(min, max T) bool
 	Extract func(pq.Value) T
+
+	helper DictionaryPredicateHelper
 }
 
 var _ Predicate = (*GenericPredicate[int64])(nil)
 
 func NewGenericPredicate[T any](fn func(T) bool, rangeFn func(T, T) bool, extract func(pq.Value) T) *GenericPredicate[T] {
-	return &GenericPredicate[T]{fn, rangeFn, extract}
+	return &GenericPredicate[T]{Fn: fn, RangeFn: rangeFn, Extract: extract}
 }
 
 func (p *GenericPredicate[T]) KeepColumnChunk(c pq.ColumnChunk) bool {
+	p.helper.setNewRowGroup()
+
 	if p.RangeFn == nil {
 		return true
 	}
@@ -305,22 +324,24 @@ func (p *GenericPredicate[T]) KeepPage(page pq.Page) bool {
 		}
 	}
 
-	// If a dictionary column then ensure at least one matching
-	// value exists in the dictionary
-	dict := page.Dictionary()
-	if dict != nil && dict.Len() > 0 {
-		len := dict.Len()
-		for i := 0; i < len; i++ {
-			if p.KeepValue(dict.Index(int32(i))) {
-				return true
+	return p.helper.keepPage(func() bool {
+		// If a dictionary column then ensure at least one matching
+		// value exists in the dictionary
+		dict := page.Dictionary()
+		if dict != nil && dict.Len() > 0 {
+			len := dict.Len()
+			for i := 0; i < len; i++ {
+				if p.KeepValue(dict.Index(int32(i))) {
+					return true
+				}
 			}
+
+			// No values matched
+			return false
 		}
 
-		// No values matched
-		return false
-	}
-
-	return true
+		return true
+	})
 }
 
 func (p *GenericPredicate[T]) KeepValue(v pq.Value) bool {
@@ -484,4 +505,29 @@ func (p *InstrumentedPredicate) KeepValue(v pq.Value) bool {
 	}
 
 	return false
+}
+
+// DictionaryPredicateHelper is a helper for a predicate that uses a dictionary
+// for filtering. it implements the Predicate interface and can be used
+// to simply dictionary handling.
+//
+// There is one dictionary per ColumnChunk/RowGroup, but it is not accessible in
+// KeepColumnChunk. This predicate saves the result of KeepPage and uses it for
+// all pages in the row group
+type DictionaryPredicateHelper struct {
+	newRowGroup         bool
+	keepPagesInRowGroup bool
+}
+
+func (d *DictionaryPredicateHelper) setNewRowGroup() {
+	d.newRowGroup = true
+}
+
+func (d *DictionaryPredicateHelper) keepPage(fn func() bool) bool {
+	if !d.newRowGroup {
+		return d.keepPagesInRowGroup
+	}
+
+	d.keepPagesInRowGroup = fn()
+	return d.keepPagesInRowGroup
 }