**What this PR does**:
Improves parquet-query's handling of dictionaries in predicates. Currently we test the dictionary for every page. However, the dictionary is per row group and not per page. Therefore, for row groups that contained multiple pages, we were repeatedly and unnecessarily checking the dictionary.

~~There is also a cardinality question. If every value in the row group is unique then testing the dictionary isn't worth it. We will test n/2 values in the dictionary until we get to the one that suggests we should keep the column and then go back and test n values for whatever the initial condition was. In this case, especially since we've already paid the cost of pulling everything from object storage, it would likely be faster to simply test the values directly.~~
Although this feels very true, benchmarks are showing otherwise. See below.

**Checklist**[
- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`