@joe-elliott 2022-12-14
I clearly have some broken logic in here due to the failed tests. Will dig into it tomorrow.