@joe-elliott 2022-12-15
I think this is in a place to be reviewed again. Benchmarks caught a nasty performance regression due to the `OrPredicate` not calling all subpredicates. I'm tempted to add logic like this instead of using the `newRowGroup` boolean, but I'm concerned about it's dependence on the underlying implementation of page and dictionary.

```
	curDict := page.Dictionary()
	if curDict == d.prevDict {
		return d.keepPagesInRowGroup
	}
	d.prevDict = curDict
```

I also settled on the value 1.2 for the `fullScanThreshold`. That means if the dictionary has ~80% of the values of the actual column it will default to just scanning the column. This is not tuned.

Benchmarks performed on a 700MB level 0 block from ops:
```
name                                     old time/op    new time/op    delta
BackendBlockSearchTraces/noMatch-8         36.8ms ±12%    39.1ms ± 2%    +6.07%  (p=0.012 n=10+8)
BackendBlockSearchTraces/partialMatch-8     3.50s ± 6%     3.36s ± 9%      ~     (p=0.105 n=8+8)
BackendBlockSearchTraces/service.name-8    3.23ms ± 5%    3.25ms ± 7%      ~     (p=0.842 n=10+9)
BackendBlockSearchTags-8                   53.2ms ± 4%    13.8ms ± 3%   -74.15%  (p=0.000 n=10+10)
BackendBlockSearchTagValues/foo-8          41.7ms ± 2%    39.6ms ± 3%    -4.98%  (p=0.000 n=9+9)
BackendBlockSearchTagValues/http.url-8      145ms ± 5%     149ms ± 3%      ~     (p=0.123 n=10+10)
BackendBlockTraceQL/noMatch-8              38.2ms ±10%    35.8ms ± 2%    -6.08%  (p=0.042 n=9+7)
BackendBlockTraceQL/partialMatch-8          233ms ±10%      79ms ±16%   -66.18%  (p=0.000 n=10+9)
BackendBlockTraceQL/service.name-8         2.79ms ± 2%    2.97ms ± 4%    +6.61%  (p=0.000 n=9+9)

name                                     old speed      new speed      delta
BackendBlockSearchTraces/noMatch-8        672MB/s ±11%   632MB/s ± 2%    -6.03%  (p=0.012 n=10+8)
BackendBlockSearchTraces/partialMatch-8  16.9MB/s ± 6%  17.7MB/s ± 9%      ~     (p=0.089 n=8+8)
BackendBlockSearchTraces/service.name-8  1.63GB/s ± 5%  1.62GB/s ± 6%      ~     (p=0.780 n=10+9)
BackendBlockTraceQL/noMatch-8             512MB/s ±10%   543MB/s ± 2%    +6.11%  (p=0.042 n=9+7)
BackendBlockTraceQL/partialMatch-8        125MB/s ± 9%   371MB/s ±16%  +196.73%  (p=0.000 n=10+9)
BackendBlockTraceQL/service.name-8        207MB/s ± 2%   194MB/s ± 4%    -6.19%  (p=0.000 n=9+9)

name                                     old bytes/op   new bytes/op   delta
BackendBlockSearchTraces/noMatch-8         24.7MB ± 0%    24.7MB ± 0%      ~     (all equal)
BackendBlockSearchTraces/partialMatch-8    59.2MB ± 0%    59.2MB ± 0%      ~     (all equal)
BackendBlockSearchTraces/service.name-8    5.27MB ± 0%    5.26MB ± 0%    -0.11%  (p=0.028 n=10+9)

name                                     old alloc/op   new alloc/op   delta
BackendBlockSearchTraces/noMatch-8          918kB ±54%    1164kB ±43%      ~     (p=0.123 n=10+10)
BackendBlockSearchTraces/partialMatch-8     322MB ± 1%     322MB ± 2%      ~     (p=0.436 n=10+10)
BackendBlockSearchTraces/service.name-8    1.54MB ± 4%    1.57MB ± 6%      ~     (p=0.190 n=10+10)
BackendBlockSearchTags-8                   13.7MB ±15%    11.8MB ± 2%   -13.73%  (p=0.000 n=10+10)
BackendBlockSearchTagValues/foo-8          1.10MB ±42%    0.92MB ±42%      ~     (p=0.113 n=10+9)
BackendBlockSearchTagValues/http.url-8     87.8MB ± 3%    87.8MB ± 3%      ~     (p=0.796 n=10+10)
BackendBlockTraceQL/noMatch-8               815kB ±37%     955kB ±34%      ~     (p=0.130 n=8+8)
BackendBlockTraceQL/partialMatch-8         55.6MB ± 1%    54.6MB ± 1%    -1.90%  (p=0.000 n=8+9)
BackendBlockTraceQL/service.name-8         76.5kB ± 0%    74.9kB ± 5%      ~     (p=0.515 n=8+10)

name                                     old allocs/op  new allocs/op  delta
BackendBlockSearchTraces/noMatch-8          4.54k ± 0%     4.54k ± 0%      ~     (p=0.119 n=10+10)
BackendBlockSearchTraces/partialMatch-8     17.9M ± 0%     17.9M ± 0%    -0.00%  (p=0.006 n=10+8)
BackendBlockSearchTraces/service.name-8     28.7k ± 0%     28.7k ± 0%      ~     (p=0.953 n=9+10)
BackendBlockSearchTags-8                    6.02k ± 0%     1.94k ± 0%   -67.69%  (p=0.000 n=10+10)
BackendBlockSearchTagValues/foo-8           4.52k ± 0%     4.52k ± 0%    -0.06%  (p=0.017 n=10+8)
BackendBlockSearchTagValues/http.url-8      2.36k ± 3%     2.41k ± 0%      ~     (p=0.250 n=10+9)
BackendBlockTraceQL/noMatch-8               3.14k ± 0%     3.14k ± 0%      ~     (p=0.481 n=8+7)
BackendBlockTraceQL/partialMatch-8          36.9k ± 0%     36.9k ± 0%    -0.01%  (p=0.003 n=8+10)
BackendBlockTraceQL/service.name-8            592 ± 0%       592 ± 0%      ~     (all equal)

name                                     old MB_io/op   new MB_io/op   delta
BackendBlockTraceQL/noMatch-8                19.5 ± 0%      19.5 ± 0%      ~     (all equal)
BackendBlockTraceQL/partialMatch-8           29.1 ± 0%      29.1 ± 0%      ~     (all equal)
BackendBlockTraceQL/service.name-8           0.58 ± 0%      0.58 ± 0%      ~     (all equal)
```