@joe-elliott 2022-12-15
Removing the "skip dictionary" threshold. As clever as I think I'm being it's apparently always better to just use the dictionary. Benchmarks from main:

```
name                                     old time/op    new time/op    delta
BackendBlockSearchTraces/noMatch-8         36.8ms ±12%    30.7ms ± 1%   -16.72%  (p=0.000 n=10+9)
BackendBlockSearchTraces/partialMatch-8     3.50s ± 6%     3.36s ±38%      ~     (p=0.829 n=8+10)
BackendBlockSearchTraces/service.name-8    3.23ms ± 5%    2.70ms ± 3%   -16.66%  (p=0.000 n=10+10)
BackendBlockSearchTags-8                   53.2ms ± 4%    10.4ms ±10%   -80.51%  (p=0.000 n=10+10)
BackendBlockSearchTagValues/foo-8          41.7ms ± 2%    36.9ms ± 6%   -11.40%  (p=0.000 n=9+10)
BackendBlockSearchTagValues/http.url-8      145ms ± 5%     128ms ± 2%   -12.11%  (p=0.000 n=10+10)
BackendBlockTraceQL/noMatch-8              38.2ms ±10%    30.7ms ± 3%   -19.52%  (p=0.000 n=9+10)
BackendBlockTraceQL/partialMatch-8          233ms ±10%      67ms ±21%   -71.12%  (p=0.000 n=10+9)
BackendBlockTraceQL/service.name-8         2.79ms ± 2%    2.66ms ±12%    -4.67%  (p=0.004 n=9+9)

name                                     old speed      new speed      delta
BackendBlockSearchTraces/noMatch-8        672MB/s ±11%   804MB/s ± 1%   +19.66%  (p=0.000 n=10+9)
BackendBlockSearchTraces/partialMatch-8  16.9MB/s ± 6%  18.6MB/s ±46%      ~     (p=0.829 n=8+10)
BackendBlockSearchTraces/service.name-8  1.63GB/s ± 5%  1.95GB/s ± 3%   +19.73%  (p=0.000 n=10+10)
BackendBlockTraceQL/noMatch-8             512MB/s ±10%   632MB/s ± 2%   +23.50%  (p=0.000 n=9+9)
BackendBlockTraceQL/partialMatch-8        125MB/s ± 9%   437MB/s ±18%  +248.96%  (p=0.000 n=10+9)
BackendBlockTraceQL/service.name-8        207MB/s ± 2%   217MB/s ±11%    +5.10%  (p=0.004 n=9+9)

name                                     old bytes/op   new bytes/op   delta
BackendBlockSearchTraces/noMatch-8         24.7MB ± 0%    24.7MB ± 0%      ~     (all equal)
BackendBlockSearchTraces/partialMatch-8    59.2MB ± 0%    59.2MB ± 0%      ~     (all equal)
BackendBlockSearchTraces/service.name-8    5.27MB ± 0%    5.26MB ± 0%      ~     (p=0.065 n=10+9)

name                                     old alloc/op   new alloc/op   delta
BackendBlockSearchTraces/noMatch-8          918kB ±54%     820kB ±42%      ~     (p=0.436 n=10+10)
BackendBlockSearchTraces/partialMatch-8     322MB ± 1%     322MB ± 1%      ~     (p=0.912 n=10+10)
BackendBlockSearchTraces/service.name-8    1.54MB ± 4%    1.57MB ± 3%      ~     (p=0.089 n=10+10)
BackendBlockSearchTags-8                   13.7MB ±15%    11.8MB ± 2%   -13.36%  (p=0.000 n=10+10)
BackendBlockSearchTagValues/foo-8          1.10MB ±42%    0.82MB ±48%   -25.74%  (p=0.028 n=10+9)
BackendBlockSearchTagValues/http.url-8     87.8MB ± 3%    87.7MB ± 0%      ~     (p=0.368 n=10+6)
BackendBlockTraceQL/noMatch-8               815kB ±37%     755kB ±16%      ~     (p=0.200 n=8+9)
BackendBlockTraceQL/partialMatch-8         55.6MB ± 1%    54.4MB ± 1%    -2.26%  (p=0.000 n=8+9)
BackendBlockTraceQL/service.name-8         76.5kB ± 0%    74.8kB ± 4%    -2.27%  (p=0.021 n=8+10)

name                                     old allocs/op  new allocs/op  delta
BackendBlockSearchTraces/noMatch-8          4.54k ± 0%     4.54k ± 0%      ~     (p=0.389 n=10+10)
BackendBlockSearchTraces/partialMatch-8     17.9M ± 0%     17.9M ± 0%      ~     (p=0.579 n=10+10)
BackendBlockSearchTraces/service.name-8     28.7k ± 0%     28.7k ± 0%    +0.06%  (p=0.000 n=9+10)
BackendBlockSearchTags-8                    6.02k ± 0%     1.93k ± 0%   -68.02%  (p=0.000 n=10+10)
BackendBlockSearchTagValues/foo-8           4.52k ± 0%     4.52k ± 0%      ~     (p=0.054 n=10+9)
BackendBlockSearchTagValues/http.url-8      2.36k ± 3%     2.32k ± 1%      ~     (p=0.246 n=10+8)
BackendBlockTraceQL/noMatch-8               3.14k ± 0%     3.14k ± 0%      ~     (p=0.314 n=8+10)
BackendBlockTraceQL/partialMatch-8          36.9k ± 0%     36.9k ± 0%    -0.02%  (p=0.001 n=8+8)
BackendBlockTraceQL/service.name-8            592 ± 0%       592 ± 0%      ~     (all equal)

name                                     old MB_io/op   new MB_io/op   delta
BackendBlockTraceQL/noMatch-8                19.5 ± 0%      19.5 ± 0%      ~     (all equal)
BackendBlockTraceQL/partialMatch-8           29.1 ± 0%      29.1 ± 0%      ~     (all equal)
BackendBlockTraceQL/service.name-8           0.58 ± 0%      0.58 ± 0%      ~     (all equal)
```